refactor(worker): extract workers/storage.mjs IO layer from api.mjs (#510)
Fourth slice of the api.mjs de-monolith — the storage/IO foundation layer.
Move the 10 artifact-read / KV / timeout / logging helpers into a new leaf
module workers/storage.mjs: logEvent, r2TimeoutMs, d1TimeoutMs, withTimeout,
readArtifact, readAsset, readR2, latestR2Key, latestPointer, readHealthKv
(plus the DEFAULT_R2/D1_TIMEOUT_MS constants).

Pure restructuring — NO behaviour change. storage.mjs imports only the
artifact-tier contract (../src/artifact-storage.mjs) and METAGRAPH_LATEST_KEY
(./config.mjs); it calls nothing back into api.mjs, so it is a clean leaf that
handlers and the response builders can share without a cycle. api.mjs now
imports the 6 it still uses (the other 4 are intra-module); dropped the
now-unused artifact-storage + METAGRAPH_LATEST_KEY imports from api.mjs.

This unblocks the next handler-group splits, which all read artifacts/health
through these helpers.

api.mjs: 3248 -> 3064 LOC (3570 at the start of #510). Verified by dependency
analysis: storage calls no api.mjs fn; none of the 10 are imported by tests
(exercised via handleRequest), so no re-exports.

Verified: 231 worker tests pass across 7 suites (worker-runtime, api-coverage,
subnet-overview, raw-artifact-freshness, health-prober, webhooks-worker,
ai-search), worker:test (esbuild) green, validate:contract-drift clean, lint
clean. No logic diff.
